requirements.txt
numpy==1.26.4
Pillow==10.3.0
streamlit==1.32.0
torch==2.2.2
torchvision==0.17.2
requests==2.31.0
git+https://github.com/facebookresearch/segment-anything.git
